all: Fail if not invoked from within the expected tool environment
Without having entered the corresponding tool's environment these
scripts are of little use. Thus we check if a certain unique enviroment
variable is set and then assume that we are in the correct environment.

This avoid cryptic error/warning messages the user might not relate to
not being in the right environment.

Signed-off-by: Joachim Foerster <joachim.foerster@missinglinkelectronics.com>