petalinux.mk: Fix, support multiple .hdf files in one directory
petalinux-config expects to get the directoy of the desired HDF file. Thus, to
keep things unambiguous, only one .hdf file can be placed in a such
directory. Otherwise it is unclear which .hdf is picked up by PetaLinux.

This patch creates a dedicated directory for the chosen .hdf file and symlinks
the chosen .hdf file into that directory. PetaLinux is then pointed to that
directory:

build/psmake/tmphdf/<hdf-basename>/<hdf-basename>

Signed-off-by: Joachim Foerster <joachim.foerster@missinglinkelectronics.com>