vitis: Implement app SRC option via importsource command
The app SRC option does not accept folders, but only file entries;
therefore while one can use the Make wildcard function, one will end up
with a flat and therefore distorted file hierarchy in the app project.
It is therefore not possible to properly organize sources of different
app projects in separate folders. In addition, one cannot import a
third-party codebase.

XSDK already has the importsource Tcl command for a long time, but it
only allowed to copy the source files, but was not able to link to them.
Vitis 2020.1 adds the -soft-link option to the importsource command that
accepts single files as well as folders. Therefore this commit replaces
the ln-based SRC option implementation by the importsource Tcl command
(in line with our policy of PSMake being a thin wrapper around the Tcl
commands whenever possible).

No symlinks are created in the src folder anymore, but Vitis adds XML
link elements to the .project file in the app project instead. Vitis
2020.1 and 2020.2 fails to deal with relative paths as arguments for the
importsource command; therefore one has to convert every source path to
an absolute path. Therefore, whenever source files are moved, one has to
rebuild the project workspace.

Signed-off-by: Stefan Wiehler <stefan.wiehler@missinglinkelectronics.com>