vitis: Harmonize usage of term "hardware description file"
The term "hardware description file" is used to subsume all different
file format implementations (HDF, DSA and XSA) from now on. The
abbreviation "HDF" shall only be used as variable name; in the
documentation or in comments the long format shall always be used to
avoid confusion with the HDF file format.

Xilinx tends to use this terminology as well (see, for example, the
PetaLinux Reference Guides for 2019.1 and 2019.2), but is not always
consequent with it. petalinux.mk already decided to stick to HDF when
transitioning from the HDF file format to XSA.

Since Vivado/PetaLinux/Vitis 2019.2, XSA is the default file format;
therefore do not mention the HDF format anymore.

Remove all references to DSA files, since this format seems to have been
rather short-lived and has never been properly tested.

Signed-off-by: Stefan Wiehler <stefan.wiehler@missinglinkelectronics.com>