vitis.mk: Add target to update platform if xsa is newer
This commit adds a new target to update the hardware platform if the specified
.xsa (`HDF` variable) is newer. Subsequently, a new stamp-file is introduced
to track the modification date. This file is (mostly) independent of Vitis
because Vitis also touches certain files during the build process of other
targets. This also fixes the bug that Vitis did not update e.g. `xparameters.h`
when a new `.xsa` was passed. Afterwards, due to the corrected dependencies,
the FSBL and apps are rebuilt using the new Makefile-targets.

Example:

    make HDF=${PATHTONEWXSA} ${YOURPLATFORMNAME}_updatehw

Signed-off-by: Frederik Pfautsch <frederik.pfautsch@missinglinkelectronics.com>