common.mk: Rewrite bitstream extraction rules to be compatible with non-ZynqMP archs
Currently, if `_BIF_ATTR` is equal to "destination_device=pl", `_BIF_FILE` can
be omitted and the bitstream is extracted from the `.xsa`. However, `destination_device`
is incompatible with Zynq and Versal architectures. Thus, there is no way of
knowing which component is the bitstream. Instead of relying on complicated
mechanics to detect the appropriate component and fill the right variable,
`$(BIT)` is simply conditionally pre-set to the right value. The user can just
set `_BIF_FILE` to `$(BIT)` and leave `_BIF_ATTR` empty.

The old behaviour of being able to leave `_BIF_FILE` empty with `_BIF_ATTR`
containing "destination_device=pl" is kept for compatibility reasons.

Signed-off-by: Frederik Pfautsch <frederik.pfautsch@missinglinkelectronics.com>