vitis.mk: Add `clean` target to applications
The README.md specifies that appliations have a _clean and _distclean target for
cleaning. Only _distclean is implemented so far, using Vitis' `app remove`
command. Due to the recently introduced integration of the applications' own
Makefile, this commit addresses the issue of the missing _clean target by
using the clean target of said Makefile without removing the app completely
(and thus having to re-add it again later).

Signed-off-by: Frederik Pfautsch <frederik.pfautsch@missinglinkelectronics.com>