vitis.mk: Reorder bitstream download *after* FSBL execution
This commit adjusts the JTAG commands to boot applications directly, such that
the bitstream is downloaded and programmed *after* the FSBL has run to ensure
similar behaviour to BOOT.BIN. This also enables e.g. the FSBL to program
external chips necessary for proper operation of the PL at initialization time
(e.g. transceivers which need to have a proper clock).

Signed-off-by: Frederik Pfautsch <frederik.pfautsch@missinglinkelectronics.com>